+++ src/ralph/models/xapi/video/fields/contexts.py
@@ -1,24 +1,22 @@
 """Video xAPI events context fields definitions"""
 
 from typing import Literal, Optional
+from uuid import UUID
 
-from pydantic import UUID4, Field
+from pydantic import Field, NonNegativeFloat
 
 from ...base import BaseModelWithConfig
 from ...fields.contexts import ContextActivitiesContextField, ContextField
 from ..constants import (
     VIDEO_CONTEXT_CATEGORY,
-    VIDEO_EXTENSION_CC_SUBTITLE_ENABLED,
+    VIDEO_EXTENSION_CC_ENABLED,
     VIDEO_EXTENSION_CC_SUBTITLE_LANG,
     VIDEO_EXTENSION_COMPLETION_THRESHOLD,
-    VIDEO_EXTENSION_FRAME_RATE,
     VIDEO_EXTENSION_FULL_SCREEN,
     VIDEO_EXTENSION_LENGTH,
-    VIDEO_EXTENSION_QUALITY,
     VIDEO_EXTENSION_SCREEN_SIZE,
     VIDEO_EXTENSION_SESSION_ID,
     VIDEO_EXTENSION_SPEED,
-    VIDEO_EXTENSION_TRACK,
     VIDEO_EXTENSION_USER_AGENT,
     VIDEO_EXTENSION_VIDEO_PLAYBACK_SIZE,
     VIDEO_EXTENSION_VOLUME,
@@ -52,10 +50,10 @@ class VideoContextExtensionsField(BaseModelWithConfig):
     """Represents the common context.extensions field for video xAPI statement.
 
     Attributes:
-        session (uuid4): Consists of the ID of the active session.
+        session (uuid): Consists of the ID of the active session.
     """
 
-    session: Optional[UUID4] = Field(alias=VIDEO_EXTENSION_SESSION_ID)
+    session_id: Optional[UUID] = Field(alias=VIDEO_EXTENSION_SESSION_ID)
 
 
 class VideoInitializedContextExtensionsField(VideoContextExtensionsField):
@@ -67,15 +65,12 @@ class VideoInitializedContextExtensionsField(VideoContextExtensionsField):
             enabled.
         ccSubtitleLanguage (str): Consists of the language of subtitle or closed
             captioning.
-        frameRate (float): Consists of the frame rate or frames per second of a video.
         fullScreen (bool): Indicates whether the video is played in full screen mode.
-        quality (str): Consists of the video resolution or quality.
         screenSize (str): Consists of the device playback screen size or the maximum
             available screen size for Video playback.
         videoPlaybackSize (str): Consists of the size in Width x Height of the video as
             viewed by the user.
         speed (str): Consists of the play back speed.
-        track (str): Consists of the name of the audio track in a media object.
         userAgent (str): Consists of the User Agent string of the browser,
             if the video is launched in browser.
         volume (int): Consists of the volume of the video.
@@ -83,16 +78,13 @@ class VideoInitializedContextExtensionsField(VideoContextExtensionsField):
             consumed to trigger a completion.
     """
 
-    length: Optional[float] = Field(alias=VIDEO_EXTENSION_LENGTH)
-    ccSubtitleEnabled: Optional[bool] = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_ENABLED)
-    ccSubtitleLanguage: Optional[str] = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_LANG)
-    frameRate: Optional[float] = Field(alias=VIDEO_EXTENSION_FRAME_RATE)
+    length: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_LENGTH)
+    ccSubtitleEnabled: Optional[bool] = Field(alias=VIDEO_EXTENSION_CC_ENABLED)
+    ccSubtitleLang: Optional[str] = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_LANG)
     fullScreen: Optional[bool] = Field(alias=VIDEO_EXTENSION_FULL_SCREEN)
-    quality: Optional[str] = Field(alias=VIDEO_EXTENSION_QUALITY)
     screenSize: Optional[str] = Field(alias=VIDEO_EXTENSION_SCREEN_SIZE)
     videoPlaybackSize: Optional[str] = Field(alias=VIDEO_EXTENSION_VIDEO_PLAYBACK_SIZE)
     speed: Optional[str] = Field(alias=VIDEO_EXTENSION_SPEED)
-    track: Optional[str] = Field(alias=VIDEO_EXTENSION_TRACK)
     userAgent: Optional[str] = Field(alias=VIDEO_EXTENSION_USER_AGENT)
     volume: Optional[int] = Field(alias=VIDEO_EXTENSION_VOLUME)
     completionThreshold: Optional[float] = Field(
@@ -111,40 +103,48 @@ class VideoBrowsingContextExtensionsField(VideoContextExtensionsField):
         length (float): Consists of the length of the video.
     """
 
-    length: Optional[float] = Field(alias=VIDEO_EXTENSION_LENGTH)
+    length: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_LENGTH)
     completionThreshold: Optional[float] = Field(
         alias=VIDEO_EXTENSION_COMPLETION_THRESHOLD
     )
 
 
-class VideoInteractedContextExtensionsField(VideoContextExtensionsField):
+class VideoEnableClosedCaptioningContextExtensionsField(VideoContextExtensionsField):
     """Represents the context.extensions field for video `interacted` xAPI statement.
 
     Attributes:
-        length (float): Consists of the length of the video.
-        ccSubtitleEnabled (bool): Indicates whether subtitle or closed captioning is
-            enabled.
         ccSubtitleLanguage (str): Consists of the language of subtitle or closed
             captioning.
-        frameRate (float): Consists of the frame rate or frames per second of a video.
-        fullScreen (bool): Indicates whether the video is played in full screen mode.
-        quality (str): Consists of the video resolution or quality.
-        videoPlaybackSize (str): Consists of the size in Width x Height of the video as
-            viewed by the user.
-        speed (str): Consists of the play back speed.
-        track (str): Consists of the name of the audio track in a media object.
+    """
+
+    ccSubtitleLanguage: str = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_LANG)
+
+
+class VideoVolumeChangeInteractionContextExtensionsField(VideoContextExtensionsField):
+    """Represents the context.extensions field for video volume change interaction xAPI
+    statement.
+
+    Attributes:
         volume (int): Consists of the volume of the video.
     """
 
-    ccSubtitleEnabled: Optional[bool] = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_ENABLED)
-    ccSubtitleLanguage: Optional[str] = Field(alias=VIDEO_EXTENSION_CC_SUBTITLE_LANG)
-    frameRate: Optional[float] = Field(alias=VIDEO_EXTENSION_FRAME_RATE)
-    fullScreen: Optional[bool] = Field(alias=VIDEO_EXTENSION_FULL_SCREEN)
-    quality: Optional[str] = Field(alias=VIDEO_EXTENSION_QUALITY)
-    videoPlaybackSize: Optional[str] = Field(alias=VIDEO_EXTENSION_VIDEO_PLAYBACK_SIZE)
-    speed: Optional[str] = Field(alias=VIDEO_EXTENSION_SPEED)
-    track: Optional[str] = Field(alias=VIDEO_EXTENSION_TRACK)
-    volume: Optional[int] = Field(alias=VIDEO_EXTENSION_VOLUME)
+    volume: int = Field(alias=VIDEO_EXTENSION_VOLUME)
+
+
+class VideoScreenChangeInteractionContextExtensionsField(VideoContextExtensionsField):
+    """Represents the context.extensions field for video screen change interaction xAPI
+    statement.
+
+    Attributes:
+        fullScreen (bool): Indicates whether the video is played in full screen mode.
+        screenSize (str): Expresses the total available screen size for Video playback.
+        videoPlaybackSize (str): Consists of the size in Width x Height of the video as
+            viewed by the user.
+    """
+
+    fullScreen: bool = Field(alias=VIDEO_EXTENSION_FULL_SCREEN)
+    screenSize: str = Field(alias=VIDEO_EXTENSION_SCREEN_SIZE)
+    videoPlaybackSize: str = Field(alias=VIDEO_EXTENSION_VIDEO_PLAYBACK_SIZE)
 
 
 class VideoInitializedContextField(BaseVideoContextField):
@@ -152,10 +152,9 @@ class VideoInitializedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoInitializedContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
-    extensions: Optional[VideoInitializedContextExtensionsField]
+    extensions: VideoInitializedContextExtensionsField
 
 
 class VideoPlayedContextField(BaseVideoContextField):
@@ -163,7 +162,6 @@ class VideoPlayedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
     extensions: Optional[VideoContextExtensionsField]
@@ -174,10 +172,9 @@ class VideoPausedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoBrowsingContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
-    extensions: Optional[VideoBrowsingContextExtensionsField]
+    extensions: VideoBrowsingContextExtensionsField
 
 
 class VideoSeekedContextField(BaseVideoContextField):
@@ -185,7 +182,6 @@ class VideoSeekedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
     extensions: Optional[VideoContextExtensionsField]
@@ -196,10 +192,9 @@ class VideoCompletedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoBrowsingContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
-    extensions: Optional[VideoBrowsingContextExtensionsField]
+    extensions: VideoBrowsingContextExtensionsField
 
 
 class VideoTerminatedContextField(BaseVideoContextField):
@@ -207,18 +202,38 @@ class VideoTerminatedContextField(BaseVideoContextField):
 
     Attributes:
         extensions (dict): See VideoBrowsingContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
     """
 
-    extensions: Optional[VideoBrowsingContextExtensionsField]
+    extensions: VideoBrowsingContextExtensionsField
 
 
-class VideoInteractedContextField(BaseVideoContextField):
-    """Represents the context field for video `interacted` xAPI statement.
+class VideoEnableClosedCaptioningContextField(BaseVideoContextField):
+    """Represents the context field for video enable closed captioning xAPI statement.
 
     Attributes:
-        extensions (dict): See VideoInteractedContextExtensionsField.
-        contextActivities (dict): See VideoContextActivitiesField.
+        extensions (dict): See VideoEnableClosedCaptioningContextExtensionsField.
     """
 
-    extensions: Optional[VideoInteractedContextExtensionsField]
+    extensions: VideoEnableClosedCaptioningContextExtensionsField
+
+
+class VideoVolumeChangeInteractionContextField(BaseVideoContextField):
+    """Represents the context field for video volume change interaction xAPI
+    statement.
+
+    Attributes:
+        extensions (dict): See VideoVolumeChangeInteractionContextExtensionsField.
+    """
+
+    extensions: VideoVolumeChangeInteractionContextExtensionsField
+
+
+class VideoScreenChangeInteractionContextField(BaseVideoContextField):
+    """Represents the context field for video screen change interaction xAPI
+    statement.
+
+    Attributes:
+        extensions (dict): See VideoScreenChangeInteractionContextExtensionsField.
+    """
+
+    extensions: VideoScreenChangeInteractionContextExtensionsField