+++ src/ralph/models/xapi/video/fields/results.py
@@ -3,12 +3,12 @@
 from datetime import timedelta
 from typing import Literal, Optional
 
-from pydantic import Field
+from pydantic import Field, NonNegativeFloat
 
 from ...base import BaseModelWithConfig
 from ...fields.results import ResultField
 from ..constants import (
-    VIDEO_EXTENSION_LENGTH,
+    VIDEO_EXTENSION_CC_ENABLED,
     VIDEO_EXTENSION_PLAYED_SEGMENTS,
     VIDEO_EXTENSION_PROGRESS,
     VIDEO_EXTENSION_TIME,
@@ -17,28 +17,28 @@ from ..constants import (
 )
 
 
-class VideoActionResultExtensionsField(BaseModelWithConfig):
-    """Represents the result.extensions field for video `played` xAPI statement.
-
-    Attributes:
-        time (float): Consists of the video time code when the event was emitted.
-    """
-
-    time: Optional[float] = Field(alias=VIDEO_EXTENSION_TIME)
-
-
-class VideoPausedResultExtensionsField(VideoActionResultExtensionsField):
+class VideoResultExtensionsField(BaseModelWithConfig):
     """Represents the result.extensions field for video `played` xAPI statement.
 
     Attributes:
         playedSegments (str): Consists of parts of the video the actor watched during
             current registration in chronological order (for example,
             "0[.]5[,]12[.]22[,]15[.]55[,]55[.]99.33[,]99.33").
+        time (float): Consists of the video time code when the event was emitted.
+    """
+
+    time: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_TIME)
+    playedSegments: Optional[str] = Field(alias=VIDEO_EXTENSION_PLAYED_SEGMENTS)
+
+
+class VideoPausedResultExtensionsField(VideoResultExtensionsField):
+    """Represents the result.extensions field for video `paused` xAPI statement.
+
+    Attributes:
         progress (float): Consists of the ratio of media consumed by the actor.
     """
 
-    playedSegments: Optional[str] = Field(alias=VIDEO_EXTENSION_PLAYED_SEGMENTS)
-    progress: Optional[float] = Field(alias=VIDEO_EXTENSION_PROGRESS)
+    progress: Optional[NonNegativeFloat] = Field(alias=VIDEO_EXTENSION_PROGRESS)
 
     class Config:  # pylint: disable=missing-class-docstring
         min_anystr_length = 0
@@ -52,49 +52,50 @@ class VideoSeekedResultExtensionsField(BaseModelWithConfig):
             media object during a seek operation.
         timeTo (float): Consists of the point in time the actor changed to in a media
             object during a seek operation.
-        length (float): Consists of the actual length of the media in seconds.
-        playedSegments (str): Consists of parts of the video the actor watched during
-            current registration in chronological order.
-        progress (float): Consists of the percentage of media consumed by the actor.
     """
 
-    timeFrom: Optional[float] = Field(alias=VIDEO_EXTENSION_TIME_FROM)
-    timeTo: Optional[float] = Field(alias=VIDEO_EXTENSION_TIME_TO)
-    length: Optional[float] = Field(alias=VIDEO_EXTENSION_LENGTH)
-    playedSegments: Optional[str] = Field(alias=VIDEO_EXTENSION_PLAYED_SEGMENTS)
-    progress: Optional[float] = Field(alias=VIDEO_EXTENSION_PROGRESS)
+    timeFrom: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_TIME_FROM)
+    timeTo: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_TIME_TO)
 
     class Config:  # pylint: disable=missing-class-docstring
         min_anystr_length = 0
 
 
-class VideoCompletedResultExtensionsField(VideoActionResultExtensionsField):
+class VideoCompletedResultExtensionsField(VideoResultExtensionsField):
     """Represents the result.extensions field for video `completed` xAPI statement.
 
     Attributes:
-        playedSegments (str): Consists of parts of the video the actor watched during
-            current registration in chronological order.
         progress (float): Consists of the percentage of media consumed by the actor.
     """
 
-    playedSegments: Optional[str] = Field(alias=VIDEO_EXTENSION_PLAYED_SEGMENTS)
-    progress: Optional[float] = Field(alias=VIDEO_EXTENSION_PROGRESS)
+    progress: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_PROGRESS)
 
     class Config:  # pylint: disable=missing-class-docstring
         min_anystr_length = 0
 
 
-class VideoTerminatedResultExtensionsField(VideoActionResultExtensionsField):
+class VideoTerminatedResultExtensionsField(VideoResultExtensionsField):
     """Represents the result.extensions field for video `terminated` xAPI statement.
 
     Attributes:
-        playedSegments (str): Consists of parts of the video the actor watched during
-            current registration in chronological order.
         progress (float): Consists of the percentage of media consumed by the actor.
     """
 
-    playedSegments: Optional[str] = Field(alias=VIDEO_EXTENSION_PLAYED_SEGMENTS)
-    progress: Optional[float] = Field(alias=VIDEO_EXTENSION_PROGRESS)
+    progress: NonNegativeFloat = Field(alias=VIDEO_EXTENSION_PROGRESS)
+
+    class Config:  # pylint: disable=missing-class-docstring
+        min_anystr_length = 0
+
+
+class VideoEnableClosedCaptioningResultExtensionsField(VideoResultExtensionsField):
+    """Represents the result.extensions field for video enable closed captioning
+    `interacted` xAPI statement.
+
+    Attributes:
+        ccEnabled (bool): Indicates whether subtitles are enabled.
+    """
+
+    ccEnabled: bool = Field(alias=VIDEO_EXTENSION_CC_ENABLED)
 
     class Config:  # pylint: disable=missing-class-docstring
         min_anystr_length = 0
@@ -104,10 +105,10 @@ class VideoPlayedResultField(ResultField):
     """Represents the result field for video `played` xAPI statement.
 
     Attributes:
-        extensions (dict): See VideoActionResultExtensionsField.
+        extensions (dict): See VideoResultExtensionsField.
     """
 
-    extensions: Optional[VideoActionResultExtensionsField]
+    extensions: VideoResultExtensionsField
 
 
 class VideoPausedResultField(ResultField):
@@ -117,7 +118,7 @@ class VideoPausedResultField(ResultField):
         extensions (dict): See VideoPausedResultExtensionsField.
     """
 
-    extensions: Optional[VideoPausedResultExtensionsField]
+    extensions: VideoPausedResultExtensionsField
 
 
 class VideoSeekedResultField(ResultField):
@@ -127,7 +128,7 @@ class VideoSeekedResultField(ResultField):
         extensions (dict): See VideoSeekedResultExtensionsField.
     """
 
-    extensions: Optional[VideoSeekedResultExtensionsField]
+    extensions: VideoSeekedResultExtensionsField
 
 
 class VideoCompletedResultField(ResultField):
@@ -140,7 +141,7 @@ class VideoCompletedResultField(ResultField):
             current registration.
     """
 
-    extensions: Optional[VideoCompletedResultExtensionsField]
+    extensions: VideoCompletedResultExtensionsField
     completion: Optional[Literal[True]]
     duration: Optional[timedelta]
 
@@ -152,14 +153,36 @@ class VideoTerminatedResultField(ResultField):
         extensions (dict): See VideoTerminatedResultExtensionsField.
     """
 
-    extensions: Optional[VideoTerminatedResultExtensionsField]
+    extensions: VideoTerminatedResultExtensionsField
 
 
-class VideoInteractedResultField(ResultField):
-    """Represents the result field for video `terminated` xAPI statement.
+class VideoEnableClosedCaptioningResultField(ResultField):
+    """Represents the result field for video enable closed captioning
+    `interacted` xAPI statement.
 
     Attributes:
-        extensions (dict): See VideoActionResultExtensionsField.
+        extensions (dict): See VideoEnableClosedCaptioningResultExtensionsField.
     """
 
-    extensions: Optional[VideoActionResultExtensionsField]
+    extensions: VideoEnableClosedCaptioningResultExtensionsField
+
+
+class VideoVolumeChangeInteractionResultField(ResultField):
+    """Represents the result field for video volume change
+    `interacted` xAPI statement.
+
+    Attributes:
+        extensions (dict): See VideoResultExtensionsField.
+    """
+
+    extensions: VideoResultExtensionsField
+
+
+class VideoScreenChangeInteractionResultField(ResultField):
+    """Represents the result field for video screen change `interacted` xAPI statement.
+
+    Attributes:
+        extensions (dict): See VideoResultExtensionsField.
+    """
+
+    extensions: VideoResultExtensionsField