+++ tests/models/edx/converters/xapi/test_server.py
@@ -18,7 +18,7 @@ from tests.fixtures.hypothesis_strategies import custom_given
 def test_models_edx_converters_xapi_server_server_event_to_xapi_convert_constant_uuid(
     uuid_namespace, event, platform_url
 ):
-    """Tests that ServerEventToPageViewed.convert returns a JSON string with a
+    """Tests that `ServerEventToPageViewed.convert` returns a JSON string with a
     constant UUID.
     """
 
@@ -39,7 +39,7 @@ def test_models_edx_converters_xapi_server_server_event_to_xapi_convert_constant
 def test_models_edx_converters_xapi_server_server_event_to_xapi_convert_with_valid_event(  # noqa
     uuid_namespace, event, platform_url
 ):
-    """Tests that converting with ServerEventToPageViewed returns the expected xAPI
+    """Tests that converting with `ServerEventToPageViewed` returns the expected xAPI
     statement.
     """
 
@@ -80,7 +80,7 @@ def test_models_edx_converters_xapi_server_server_event_to_xapi_convert_with_val
 def test_models_edx_converters_xapi_server_server_event_to_xapi_convert_with_anonymous_user(  # noqa
     uuid_namespace, event, platform_url
 ):
-    """Tests that anonymous usernames are replaced with with `anonymous`."""
+    """Tests that anonymous usernames are replaced with `anonymous`."""
 
     event.context.user_id = ""
     event_str = event.json()